requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# ML Tools
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Progress & Logging
tqdm>=4.65.0
tensorboard>=2.13.0

# API Development
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Utilities
pyyaml>=6.0
python-multipart>=0.0.6

# Tracking & Prediction
filterpy>=1.4.5
